first full connection with redis om & redis pub/sub. message header was changed

+++ DynamicBandwidth/DynamicBandwidthCommon/Classes/RedisMessageUtility.cs
@@ -21,7 +21,7 @@ namespace DynamicBandwidthCommon
             var messageHeader = new MessageHeader()
             {
                 Id        = message.Id,
-                Priority  = priority,
+                Priority  = (int)priority,
                 DataType  = dataType.ToString(),
                 DataSize  = data.Length,
                 TimeStamp = DateTime.Now.Ticks
